pyproject.toml
[tool.poetry]
name = "weather-project"
version = "0.1.0"
description = ""
authors = ["Hiba TOUDERTI <hibatouderti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
kafka-python = "^2.0.2"
cassandra-driver = "^3.29.0"
streamlit = "^1.31.1"
pyspark = "^3.5.1"
plotly = "^5.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
